requirements.txt
flask
google-generativeai
pandas
torch
sentence-transformers
Pillow